+++ src/main/java/com/mapshot/api/MapShotApiApplication.java
@@ -3,9 +3,11 @@ package com.mapshot.api;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableScheduling
 public class MapShotApiApplication {
 
     public static void main(String[] args) {
