--- src/main/java/com/mapshot/api/admin/entity/AdminEntity.java
@@ -1,28 +1,0 @@
-package com.mapshot.api.admin.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity(name = "admin")
-public class AdminEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String nickname;
-
-    private String password;
-
-}
