--- src/main/java/com/mapshot/api/admin/model/AdminRequest.java
@@ -1,15 +1,0 @@
-package com.mapshot.api.admin.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class AdminRequest {
-    private String nickname;
-    private String password;
-}
