+++ src/main/java/com/mapshot/api/domain/admin/AdminService.java
@@ -1,0 +1,51 @@
+package com.mapshot.api.domain.admin;
+
+import com.mapshot.api.infra.web.exception.ApiException;
+import com.mapshot.api.infra.web.exception.status.ErrorCode;
+import com.mapshot.api.presentation.admin.model.AdminRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+@Service
+@RequiredArgsConstructor
+public class AdminService {
+
+    private final AdminRepository adminRepository;
+    private static final String ENCRYPT_ALGORITHM = "SHA-256";
+
+
+    @Transactional(readOnly = true)
+    public void validateUser(AdminRequest request) {
+        String nickname = request.getNickname();
+        String password = encrypt(request.getPassword());
+
+        adminRepository.findByNicknameAndPassword(nickname, password)
+                .orElseThrow(() -> new ApiException(ErrorCode.NO_SUCH_USER));
+    }
+
+
+    private String encrypt(String text) {
+        try {
+            MessageDigest md = MessageDigest.getInstance(ENCRYPT_ALGORITHM);
+            md.update(text.getBytes());
+
+            return bytesToHex(md.digest());
+        } catch (NoSuchAlgorithmException e) {
+            throw new ApiException(ErrorCode.NO_SUCH_ALGORITHM);
+        }
+
+    }
+
+    private String bytesToHex(byte[] bytes) {
+        StringBuilder builder = new StringBuilder();
+        for (byte b : bytes) {
+            builder.append(String.format("%02x", b));
+        }
+        return builder.toString();
+    }
+
+}
