+++ src/main/java/com/mapshot/api/domain/notice/NoticeConst.java
@@ -1,0 +1,5 @@
+package com.mapshot.api.domain.notice;
+
+public class NoticeConst {
+    public static final int MAX_CONTENT_LENGTH = 2000;
+}
