+++ src/main/java/com/mapshot/api/domain/notice/NoticeRepository.java
@@ -1,0 +1,13 @@
+package com.mapshot.api.domain.notice;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface NoticeRepository extends JpaRepository<NoticeEntity, Long> {
+    List<NoticeEntity> findTop10ByIdLessThanOrderByIdDesc(long id);
+
+    NoticeEntity findFirstByOrderByIdDesc();
+
+}
