--- src/main/java/com/mapshot/api/image/model/StorageInner.java
@@ -1,18 +1,0 @@
-package com.mapshot.api.image.model;
-
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class StorageInner {
-
-    private String uuid;
-    private byte[] imageByte;
-    private LocalDateTime createdAt;
-}
