--- src/main/java/com/mapshot/api/image/model/StorageRequest.java
@@ -1,18 +1,0 @@
-package com.mapshot.api.image.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@ToString
-public class StorageRequest {
-
-    private String uuid;
-    private String base64EncodedImage;
-}
