--- src/main/java/com/mapshot/api/notice/consts/NoticeConst.java
@@ -1,5 +1,0 @@
-package com.mapshot.api.notice.consts;
-
-public class NoticeConst {
-    public static final int MAX_CONTENT_LENGTH = 2000;
-}
