--- src/main/java/com/mapshot/api/notice/repository/NoticeRepository.java
@@ -1,13 +1,0 @@
-package com.mapshot.api.notice.repository;
-
-
-import com.mapshot.api.notice.entity.NoticeEntity;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NoticeRepository extends JpaRepository<NoticeEntity, Long> {
-    List<NoticeEntity> findTop10ByIdLessThanOrderByIdDesc(long id);
-
-    NoticeEntity findFirstByOrderByIdDesc();
-
-}
