+++ src/main/java/com/mapshot/api/presentation/admin/model/AdminRequest.java
@@ -1,0 +1,15 @@
+package com.mapshot.api.presentation.admin.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AdminRequest {
+    private String nickname;
+    private String password;
+}
