+++ src/main/java/com/mapshot/api/presentation/map/StorageController.java
@@ -1,0 +1,45 @@
+package com.mapshot.api.presentation.map;
+
+import com.mapshot.api.domain.map.StorageService;
+import com.mapshot.api.infra.web.auth.annotation.PreAuth;
+import com.mapshot.api.infra.web.auth.enums.Accessible;
+import com.mapshot.api.presentation.map.model.StorageInner;
+import com.mapshot.api.presentation.map.model.StorageRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDateTime;
+import java.util.Base64;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/image/storage")
+@CrossOrigin(originPatterns = {"https://*.kmapshot.com", "https://kmapshot.com"})
+public class StorageController {
+
+    private final StorageService storageService;
+
+    @PreAuth(Accessible.EVERYONE)
+    @GetMapping("/{uuid}")
+    public ResponseEntity<byte[]> returnCompletedImageToUser(@PathVariable String uuid) {
+        byte[] imageResource = storageService.getImage(uuid);
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.IMAGE_JPEG)
+                .body(imageResource);
+    }
+
+    @PreAuth(Accessible.FRIENDLY_SERVER)
+    @PostMapping
+    public void saveCompletedImage(@RequestBody StorageRequest storageRequest) {
+        StorageInner storageInner = StorageInner.builder()
+                .uuid(storageRequest.getUuid())
+                .imageByte(Base64.getDecoder().decode(storageRequest.getBase64EncodedImage()))
+                .createdAt(LocalDateTime.now())
+                .build();
+
+        storageService.add(storageInner);
+    }
+}
