+++ src/main/java/com/mapshot/api/presentation/map/model/StorageInner.java
@@ -1,0 +1,19 @@
+package com.mapshot.api.presentation.map.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class StorageInner {
+
+    private String uuid;
+    private byte[] imageByte;
+    private LocalDateTime createdAt;
+}
